Latest update
I started by changing the shop to just the shop page and got the "basket" function applied to it. So now the basket is just in the shop page not on the navbar. I then went through and cleaned up the all the css/styled code: to get rid of errors/conflicts; get rid of code that wasn't being used; move code to the more relevant files.